Reduce harpguru rerenders by batching state updates
This commit was made following the realisation that having 2 seperate
state updates would trigger 2 harpguru rerenders when only 1 is
necessary.

There are a lot of failing tests as a result, but I think perhaps it's
best to save this for reference before trying to unpick it.
