+++ packages/harpguru-core/src/components/harp-guru/hooks/use-flush-batch-id-toggles/use-flush-batch-id-toggles.ts
@@ -1,35 +1,16 @@
-import { useGlobal } from 'reactn'
+import { useGlobal, useDispatch } from 'reactn'
 import { useEffect } from 'react'
-import type { HarpStrataProps } from 'harpstrata'
-import { getHarpStrata } from 'harpstrata'
-
-import { batchToggleDegreeIds } from '../../utils'
 
 export const useFlushBatchIdToggles = (): void => {
-  const [batchIdToggleBuffer, setBatchIdToggleBuffer] = useGlobal('toggleDegreeIdsBuffer')
-  const [activeHarpStrata, setActiveHarpStrata] = useGlobal('activeHarpStrata')
+  const [batchIdToggleBuffer] = useGlobal('toggleDegreeIdsBuffer')
+  const updateHarpStrataAndFlushBuffer = useDispatch(
+    'updateHarpStrataAndFlushBuffer'
+  )
 
   useEffect(() => {
     const flushBufferedToggles = setTimeout(() => {
       if (batchIdToggleBuffer.length == 0) return
-      const {
-        apparatus: { id: apparatusId },
-        pozitionId,
-        harpKeyId,
-        isActiveComplex: { activeDegreeIds },
-      } = activeHarpStrata
-      const newActiveDegreeIds = batchToggleDegreeIds(
-        activeDegreeIds,
-        batchIdToggleBuffer
-      )
-      const newHarpStrataProps: HarpStrataProps = {
-        apparatusId,
-        pozitionId,
-        harpKeyId,
-        activeIds: newActiveDegreeIds,
-      }
-      setActiveHarpStrata(getHarpStrata(newHarpStrataProps))
-      setBatchIdToggleBuffer([])
+      updateHarpStrataAndFlushBuffer()
     }, 500)
     return () => {
       clearTimeout(flushBufferedToggles)
