+++ packages/harpguru-core/src/utils/set-global-reducers/set-global-reducers.ts
@@ -1,10 +1,12 @@
 import type { State } from 'reactn/default'
 import { addReducer } from 'reactn'
 import { getHarpStrata } from 'harpstrata'
+import type { HarpStrataProps } from 'harpstrata'
 
 import { getPropsForHarpStrata } from '../get-props-for-harp-strata'
 import { getNextQuizQuestion } from '../get-next-quiz-question'
 import { DisplayModes } from '../../types'
+import { batchToggleDegreeIds } from '../../components/harp-guru/utils'
 
 import { activateHarpCell } from './utils'
 
@@ -33,6 +35,35 @@ export const setGlobalReducers = (): void => {
     const { activeHarpStrata, quizQuestion } = global
     return {
       activeHarpStrata: activateHarpCell(activeHarpStrata, quizQuestion),
+    }
+  })
+
+  addReducer('updateHarpStrataAndFlushBuffer', (global: State) => {
+    const {
+      activeHarpStrata,
+      toggleDegreeIdsBuffer: batchIdToggleBuffer,
+    } = global
+    const {
+      apparatus: { id: apparatusId },
+      pozitionId,
+      harpKeyId,
+      isActiveComplex: { activeDegreeIds },
+    } = activeHarpStrata
+    const newActiveDegreeIds = batchToggleDegreeIds(
+      activeDegreeIds,
+      batchIdToggleBuffer
+    )
+    const newHarpStrataProps: HarpStrataProps = {
+      apparatusId,
+      pozitionId,
+      harpKeyId,
+      activeIds: newActiveDegreeIds,
+    }
+    const newHarpStrata = getHarpStrata(newHarpStrataProps)
+
+    return {
+      activeHarpStrata: newHarpStrata,
+      toggleDegreeIdsBuffer: [],
     }
   })
 }
